Fix mdr_spec_base_sz which didn't include the payload size

+++ mdr/mdr.h
@@ -353,7 +353,7 @@ const struct mdr_spec *mdr_register_spec(struct mdr_def *);
 const struct mdr_spec *mdr_registry_get(uint64_t);
 void                   mdr_registry_clear();
 
-size_t                 mdr_spec_base_sz(const struct mdr_spec *, uint64_t);
+uint64_t               mdr_spec_base_sz(const struct mdr_spec *, uint64_t);
 
 #define MDR_DCV(domain, code, variant) \
     ((uint64_t)domain << 32 | (uint64_t)code << 16 | (uint64_t)variant)
